requirements.txt
anyio==3.6.2
automium==0.2.6
automium-web==0.1.1
certifi==2022.12.7
click==8.1.3
distlib==0.3.6
fastapi==0.89.1
FastAPI-SQLAlchemy==0.2.1
filelock==3.9.0
Flask==2.2.2
greenlet==2.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
legacy==0.1.6
MarkupSafe==2.1.1
pipenv==2022.12.19
platformdirs==2.6.2
psycopg2-binary==2.9.5
pydantic==1.10.4
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.45
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
Werkzeug==2.2.2
